requirements.txt
llama-index==0.12.11
llama-index-llms-groq==0.3.1
arxiv==2.1.3
llama-index-readers-web==0.3.4
scrapfly-sdk==0.8.19
docx2txt==0.8
pypdf==5.1.0
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
groq==0.15.0
cohere==5.13.8
qdrant-client==1.13.0
llama_index_llms_cohere==0.4.0
llama-index-llms-openai==0.3.13
llama-index-embeddings-cohere==0.4.0
llama-index-llms-gemini==0.4.3
llama-index-llms-together==0.3.1
torch==2.5.1
logging==0.4.9.6
